+++ app/src/main/java/eu/kanade/tachiyomi/ui/reader/model/ReaderPage.kt
@@ -1,6 +1,7 @@
 package eu.kanade.tachiyomi.ui.reader.model
 
 import eu.kanade.tachiyomi.source.model.Page
+import okhttp3.Request
 import java.io.InputStream
 
 open class ReaderPage(
@@ -11,4 +12,5 @@ open class ReaderPage(
 ) : Page(index, url, imageUrl, null) {
 
     open lateinit var chapter: ReaderChapter
+    var remoteImageRequest: (suspend () -> Request?)? = null
 }